@timvandermeij 2019-07-19
Nice work, and really good that you revisited the ideas from the original PR!